Bug 1152272 - Part 1: Refactor the AT command handler. r=hsinyi
The original AT command handler returns a replying string (amodem_send() will
send the reply for it) or returns NULL when the reply has been sent in handler.
As we're to add mutex to protect out buffer, such inconsistent handling is not
effective and could be error prone.

This patch moves all the reply handling into AT command handler to have a
consistent way to send the reply.